third_party/WebKit/LayoutTests/fast/html/imports/import-onload-with-stylesheets.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script>
window.jsTestIsAsync = true;

function importLoaded(event)
{
    shouldBeTrue("event.target.import.querySelector('link').sheet instanceof StyleSheet");
    finishJSTest();   
}
</script>
<link rel="import" href="resources/hello-css.html" onload="importLoaded(event)">
