third_party/WebKit/LayoutTests/fast/html/imports/import-onload-with-stylesheets-expected.txt

PASS event.target.import.querySelector('link').sheet instanceof StyleSheet is true
PASS successfullyParsed is true

TEST COMPLETE
